UPSTREAM: util/gitconfig: Make checkpatch non fatal in pre-commit hook
We don't block commits for failing checkpatch in gerrit, and we
shouldn't block them here.  This allows checkpatch to still run, so
users can see the issues, but won't prevent the commit.

Adds a delay if checkpatch fails so that the issues can be seen.

BUG=none
BRANCH=none
TEST=none

Change-Id: I071cc41e7c2494d62bc884e88032b5dcd85d4926
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 4a6477ed645ce369e4a7fb2728c82fb7d3e5f985
Original-Change-Id: Ibd4e8bb74e0b02825dcdf16e233a061c4bb43f50
Original-Signed-off-by: Martin Roth <martin@coreboot.org>
Original-Reviewed-on: https://review.coreboot.org/27534
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Ronald G. Minnich <rminnich@gmail.com>
Reviewed-on: https://chromium-review.googlesource.com/1145420
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>