UPSTREAM: soc/intel/skylake: permit Kconfig to set subsystem ID
This change permits the subsystem ID to be specified via Kconfig for the
devices on the SoC.

Some devices are getting zero'ed subsystem IDs because the unset (and thus
zero'ed) config options are overwriting the fsp defaults. With this change
the fsp defaults will only be overwritten by non-zero config values.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0fbfb94ada6785b1d968cf2697e6bba1f418727a
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: dfffcad66f4bad1d08c8bd9d6058c54b939422c5
Original-Change-Id: I0f7bb8e465f55e5dd6d8e0fad71b9b2a22f089dc
Original-Signed-off-by: Matt Delco <delco@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/27609
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Duncan Laurie <dlaurie@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1181369
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>