UPSTREAM: soc/intel/common/block/gpio: Add API for gpio_configure_pads_with_override
This function adds support for gpio_configure_pads_with_override
which:
1. Takes as input two GPIO tables -- base config table and override
config table
2. Configures each pad in base config by first checking if there is a
config available for the pad in override config table. If yes, then
uses the one from override config table. Else, uses the base config to
configure the pad.

This is done to allow sharing of GPIO tables across baseboard-variants
for various boards i.e. Each board can have a base config table which
is provided by the baseboard and an optional override config table
that can be provided by a variant to configure certain GPIOs
differently. It is helpful when the variant GPIO diff list is not very
huge compared to the baseboard.

BUG=b:111743717
TEST=Verified that the GPIO config for phaser is same with and without
this change.

Change-Id: Icd65b91d7142400dc52c28031d6714ad7e09b799
Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Commit-Id: 4a12a56cdf1ea073def5e62386eebac32ff59cd3
Original-Change-Id: I1c5dc72c8368957201ab53d2e8398ff861341a4c
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/27640
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Reviewed-by: Justin TerAvest <teravest@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1152155
Commit-Ready: Furquan Shaikh <furquan@chromium.org>
Tested-by: Furquan Shaikh <furquan@chromium.org>
Reviewed-by: Furquan Shaikh <furquan@chromium.org>