UPSTREAM: device/pnp_device: rewrite pnp_get_ioresource
The rewrite simplifies the code and fixes the limit on the case of IO resources
with the length of 1 byte.

This patch is inspired by the comment of Nico Huber on
Ia99b785dcd9cf56fb236ad7ade54656851f88a5e

If the ones in the mask aren't a continuous block, now a warning message gets
printed.

BUG=none
BRANCH=none
TEST=none

Change-Id: I58bf44adfb41943b461f05e200718509b1af55fb
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f2ec648fcb6f17382293cbcc69e46c485ee3b786
Original-Change-Id: I0089278f5adf65e8f084580bc5bc1eb2f77b87f7
Original-Signed-off-by: Felix Held <felix-coreboot@felixheld.de>
Original-Reviewed-on: https://review.coreboot.org/27654
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1155544
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>