UPSTREAM: mainboard: Add ASUS P8H61-M LX
Tested with GRUB 2.02 as a payload, booting Debian GNU/Linux 9.5 with
kernel 4.9. This code is based on the output of autoport.

The file `data.vbt` matches the VBT in the latest version of the vendor
firmware (version 4601).

This board works well under coreboot. A list of what works and what
doesn't can be found in the documentation part of this commit. To
summarise: the only known issues are that S3 suspend/resume doesn't
work, and that there is no automatic fan control via the super I/O.

BUG=none
BRANCH=none
TEST=none

Change-Id: I5f174828da0bea3de78dc6c64fb29e00e1a2b90d
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 921a4cfa3feebb9c793cf45626be0a58ea32a670
Original-Change-Id: I2a0579f486d3a44de2dd927fa1e76b90c3b48f62
Original-Signed-off-by: Tristan Corrick <tristan@corrick.kiwi>
Original-Reviewed-on: https://review.coreboot.org/27798
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Angel Pons <th3fanbus@gmail.com>
Original-Reviewed-by: Arthur Heymans <arthur@aheymans.xyz>
Reviewed-on: https://chromium-review.googlesource.com/1163252
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>