UPSTREAM: mb/google/poppy/variants/rammus: add rammus devicetree.cb
Use the default value for Iccmax which is specified in vr_config.c.
The AcLoadline and DcLoadline keep the poppy value. Besides, the
USB 2.0 ports located on the mainboard are set to USB2_PORT_SHORT
and the others on the daughterboard are set to USB2_PORT_LONG.
Those setting need to be fine tuned later.

BUG=b:111579386
BRANCH=Master
TEST=Build pass

Change-Id: I426548f2e5d57659e6d36ad40477818762d4799e
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 11f016063fa0d2f8ccf187de693121d16a1e09ff
Original-Change-Id: Icabfac04c94b3d480872c243d811509e274ef122
Original-Signed-off-by: Zhuohao Lee <zhuohao@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/27808
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1163494
Commit-Ready: ChromeOS CL Exonerator Bot <chromiumos-cl-exonerator@appspot.gserviceaccount.com>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>