UPSTREAM: amd/stoneyridge: Add warm reset detection
Extend the existing reset handling features in Stoney Ridge to plan for,
and recognize, warm resets.  The ColdRstDet bit is always zero on a cold
reset, and is intended as a mechanism for the BIOS to determine the type
of a reset that occurred.

Set ColdRstDet=1 after all cores have been initialized, so that any
subsequent reset may be identified as warm/cold.  A later patch will check
the value during mp_init.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ie2ef5bf48950465e53599e7f71077de508cc19d6
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 2e49cf129ac0f3326989cd1202904a3f02d1d467
Original-Change-Id: I90255918de03018c9f090bff1e56a8bda5e7365e
Original-Signed-off-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/27924
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1167950