UPSTREAM: libpayload/arch/x86: Add support for initializing the APIC
This is just the bare minimum required to initialize the APIC. I only
support xAPIC and chose not to support x2APIC. We can add that
functionality later when it's required.

I also made the exception dispatcher call apic_eoi so that the callbacks
won't forget to call it.

BUG=b:109749762
TEST=Booted grunt and verified that depthcharge continued to function
and that linux booted correctly. Also verified GDB still works.

Change-Id: I21799f21187a71f5cd9622cb5b8bb2127f9bc47f
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: ac8ebd0e73774c83bfa5cb3dd460c662000e0d19
Original-Change-Id: I420a4eadae84df088525e727b481089ef615183f
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/28241
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1226041
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>