UPSTREAM: drivers/net/atl1e: Add driver
A shortcoming of this driver is that if multiple devices with the same PCI ID
are present and don't have an eeprom, they would all get the same macadress set.
The r8168 driver deals with such cases so it should be easy to implement if
needed.

BUG=none
BRANCH=none
TEST=none

Change-Id: I8fb77f7c719216a8b67162d3c39decacc23f41b3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7f922b0f6ad9b63c06e6243b0a485b05b36befb9
Original-Change-Id: I5c32df00e25453c350a45e7f1ee6834b89c4289f
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/28265
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1282455
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>