UPSTREAM: libpayload/x86/exception: Add ability to handle user defined interrupts
I need to setup the APIC timer to fire interrupts. I would like to reuse
the existing interrupt table. So I extended it to support user defined
interrupts. I just added all 255 vectors so there wouldn't need to be
any additional build time configuration.

I'm going to deprecate exception_install_hook and remove it in a follow
up. It will be replaced with set_interrupt_handler. This way the
exception lookup does not have to manage a list of callbacks, or have to
worry about the order they are processed.

BUG=b:109749762
TEST=Wrote an interrupt handler and fired an APIC timer interrupt and
verified that vector 32 was returned.

Change-Id: Id2cf58e4760e36472b577f2df58e7ff79b6c44b3
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 80d5c195901d7fa1a442b36d697f9f0f3f063379
Original-Change-Id: Id9c2583c7c3d9be4a06a25e546e64399f2b0620c
Original-Signed-off-by: Raul E Rangel <rrangel@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/28100
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1216485
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>