UPSTREAM: x86/acpi: Add ACPI table revision function
Use a single function to set ACPI table versions. This allows us
to keep revisions synced to the correct levels for coreboot. This
is a partial fix for the bug:

FAILED [MEDIUM] SPECMADTFADTRevisions: Test 2, MADT revision is not
in sync with the FADT revision; MADT 1 expects FADT 3.0 but found 4.0
instead.

BUG=b:112476331
TEST-Run FWTS

Change-Id: I688bea32dee17ee2cd4a2923f6a4e2e4df9444a1
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 93a51766aaca50cfbcbedca780e935ce66c12d73
Original-Change-Id: Ie9a486380e72b1754677c3cdf8190e3ceff9412b
Original-Signed-off-by: Marc Jones <marcj303@gmail.com>
Original-Reviewed-on: https://review.coreboot.org/28276
Original-Reviewed-by: Martin Roth <martinroth@google.com>
Original-Reviewed-by: Marshall Dawson <marshalldawson3rd@gmail.com>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1191645