UPSTREAM: drivers/intel/gma: Do not rely on CBLV in OpRegion Mailbox3
CBLV is not kept up to date by Linux' i915. We should fix that too,
but it will likely take some years until we can always expect it to
work.

For now read the register values directly. To accomodate that we
are not the only one writing those, revise XBQC() to search for
the closest value in BRIG (instead of a lower equal one) and round
more accurately for better matches.

BUG=none
BRANCH=none
TEST=none

Change-Id: I0b768eb60931f0249ce7a3663564fd35e8b812d8
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f26a7c76873b9a647d338370a2adf706f8b3f705
Original-Change-Id: I4e2d8fa34e75463d4cf7242af3e2c67577cfa2a5
Original-Signed-off-by: Nico Huber <nico.h@gmx.de>
Original-Reviewed-on: https://review.coreboot.org/28301
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1227774
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>