UPSTREAM: nb/intel/x4x: Don't use cached settings if CPU FSB has been changed
Using the cached CPU FSB setting can simply be wrong, in which case it won't
boot. Since the selected timings also depend on the CPU FSB, it is also best to
not use cached timings at all when a change is detected.

Tested on P5QC, swapped a 1333MHz FSB to a 800MHz FSB and it uses !fast_boot
boot path.

BUG=none
BRANCH=none
TEST=none

Change-Id: I77e2ec638964dd42648552c40d70aeba65616dd1
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: b0c6cffb093d82b8a24b00e1914e5195bcc0b50d
Original-Change-Id: I12d91d0e892c15778409d7c00b27652ee52ca80c
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/28506
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1227125
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>