UPSTREAM: mb/google/octopus: Configure H1 interrupt pad using Rx level config
This change configures GPIO_63 (which is used for H1 interrupts) as Rx
Level. This ensures that the signal gets passed on to the next logic
state as is and the APIC entry can be configured to trigger interrupt
on level or edge as per the kernel driver expectation.

TEST=Verified that no H1 interrupt timeouts are seen with 100
iterations of warm and 100 iterations of cold reboot.

Change-Id: Iba32de050d66048a02bda8595d46b281bc9ec0e3
Signed-off-by: Martin Roth <martinroth@chromium.org>
Original-Commit-Id: 2d602098f9def150d1111f334a668391c25da5cf
Original-Change-Id: I7aac30300a4251d9b40276dcca7ebc6a6d814c40
Original-Signed-off-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-on: https://review.coreboot.org/28507
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/1213899