UPSTREAM: nb/intel/x4x: Program read training results to all ranks
While during the read training itself only the settings for rank 0 are used for
all ranks, the controller does use the separate settings for each rank later on.
It is unknown which register is responsible for this.
The signals are probably not generated separately and therefore need to have the
same settings for all ranks. Therefore program the results for all ranks instead
of for all populated ranks.

TESTED: Fixes DG43GT not booting with only the second DIMM slot of a channel
populated.

BUG=none
BRANCH=none
TEST=none

Change-Id: I9060ce38be00a7009c5cc9c96faed6e42c9bc473
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 8ddd7d1e5eee9e3e604624622d9b91f9b99306ff
Original-Change-Id: I7965a068ef4779847e62e966154764370c91302a
Original-Signed-off-by: Arthur Heymans <arthur@aheymans.xyz>
Original-Reviewed-on: https://review.coreboot.org/28577
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Felix Held <felix-coreboot@felixheld.de>
Reviewed-on: https://chromium-review.googlesource.com/1282450
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>