UPSTREAM: google/kukui: Notify EC that AP is in S0
We have a pin from AP to EC, called AP_IN_SLEEP_L (SRCLKENA0 on AP side,
pad R23) that is supposed to be high in S0, and low in S3 (and X/don't
care in S5).
This should be set as early as possible in bootblock.

BUG=b:113367227
TEST=make; boots and verified AP_IN_SLEEP_L GPIO is high.
BRANCH=None

Change-Id: I779d285741dc5fb3b246c50bc2eba8761b548e34
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: f89c56a54d62495804167eef40b2718cfdade185
Original-Change-Id: Icd59fa366c162e7443b8932a851e65f110f551ab
Original-Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Original-Reviewed-on: https://review.coreboot.org/28585
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Julius Werner <jwerner@google.com>
Reviewed-on: https://chromium-review.googlesource.com/1227787
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>