UPSTREAM: soc/intel/cannonlake: Fix ACPI FADT table generation
The function to fill out the FADT table exits early if the devicetree
config option to disable the legacy timer is set.  This means it never
gets to the later check for s0ix config option and so the flag to
indicate that it supports low-power idle in S0 is not set.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ib15b704c340f9830135d52c311035fbc60a00365
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 174ca43583f33400094fb40c1e689a38946bdeac
Original-Change-Id: Ia0416f21b6445f6feecb6f0301d48fdf2522b8a6
Original-Signed-off-by: Duncan Laurie <dlaurie@google.com>
Original-Reviewed-on: https://review.coreboot.org/28755
Original-Reviewed-by: Furquan Shaikh <furquan@google.com>
Original-Reviewed-by: Aaron Durbin <adurbin@chromium.org>
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Reviewed-on: https://chromium-review.googlesource.com/1251180
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>