UPSTREAM: configs: add sercon port and disable pxe serial console for apu{2,3,4,5}
To avoid mangled characters on serial output from iPXE we have to disable
serial from iPXE console. More to that to have correct serial input we
have to enable SeaBIOS SERCON option with default configuration.

The only limitation of this configs is that apu5 doesn't detect iPXE -
that platform is not for public use so it doesn't affect anyone.

BUG=none
BRANCH=none
TEST=none

Change-Id: Ife703051d380bd2c635e890d7a0536fed1e7c353
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 1c54985bb03534037318afc6c13fb6f3a5900fb8
Original-Change-Id: I124705bd691b3c8dcd9a2636b17c019d02732c5a
Original-Signed-off-by: Piotr Krl <piotr.krol@3mdeb.com>
Original-Reviewed-on: https://review.coreboot.org/28616
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Patrick Rudolph <siro@das-labor.org>
Reviewed-on: https://chromium-review.googlesource.com/1227122
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>