UPSTREAM: siemens/mc_apl1: Make the DDR memory swizzle data configurable
In preparation for a future MC Apollo Lake board which will be equipped
with LPDDR4 modules, it is necessary to make the swizzle data
configurable. Starting from the mc_apl1 baseboard, which is equipped
with DDR3L memory and therefore does not need swizzle data, the
structures are initialized with zero.

BUG=none
BRANCH=none
TEST=none

Change-Id: I132ee92d92cd1d6b6eca6d1d59c1c4c16d5398f2
Signed-off-by: Patrick Georgi <pgeorgi@google.com>
Original-Commit-Id: 7e15e87eccf8adbfdeb22dd1bca98d977c99bde1
Original-Change-Id: I4954d0a00d1d5fc28a8dda45a9fb27f98d5c3f1e
Original-Signed-off-by: Mario Scheithauer <mario.scheithauer@siemens.com>
Original-Reviewed-on: https://review.coreboot.org/28730
Original-Tested-by: build bot (Jenkins) <no-reply@coreboot.org>
Original-Reviewed-by: Werner Zeh <werner.zeh@siemens.com>
Reviewed-on: https://chromium-review.googlesource.com/1251169
Commit-Ready: Patrick Georgi <pgeorgi@chromium.org>
Tested-by: Patrick Georgi <pgeorgi@chromium.org>
Reviewed-by: Patrick Georgi <pgeorgi@chromium.org>